BugFix: POIs were not correctly being picked up. In particular scale 24 POIs in the high-style.
The problem was that the boundaries of the tiles were determined by the roads and not the POIs.

However for example in the high-style there are no roads in tile zoom level 3. Thus the bounding
boxes for the zoom level 3 tiles shrank to points, which meant that none of the POIs at tile zoom level
3 were included in Osm2GpsMid anymore. This problem was probably present in other situations too in less
severe form and might explain some of the missing POI bug reports.

However this bug was deep in the tile generation code, and I don't understand this part properly. I thus
have no idea, if this bug fix might have subtile other effects introducing new bugs. If someone could review
this code please to check it isn't doing anything stupid.