Give an error message if the style-file contains more than 126 POI or way types
As we use a byte to indicate the way and POI type we can't support more than 127 types
at the moment.

Currently we use 5 bytes per node (2 * 2 for the coordinates and one for the type) if we used more than
a byte, the size and memory usage would go up by quite a bit, so for the moment, just return an error message
that we can't support more than 127 types.