[ patch #2857694 ] Support for GeoHack in "Online Info"
patch by jkpj (with shopefully some fixes for the Southern and Western hemispheres)

The attached patch adds support for the GeoHack web service in gpsmid's
"Online Info" menu.

GeoHack offers lots of online global and local map services. This patch
makes it easy to look at an aerial image, bird's eye view, topographical
map, national land survey map (if available as it is for Finland) etc. for
the current location. Online Info opens a web browser on the device, so
internet access is required.