Patch "Show target and move center to GPS to the left by jkpj"
This is a patch to fulfil feature request
https://sourceforge.net/tracker/?func=detail&aid=2856810&group_id=192084&at
id=939977

This has passed only cursory testing with microemulator and touchscreen UI
coding is new to me so glance through to check I don't break anything.

feature request was:

"I request the touch screen "center to GPS" (if I understand it right,
it's
the vertical bar to the right of the + / - zoom buttons) to the left side
of the zoom buttons. I have calibrated the touchscreen so it seems to work
fine for me otherwise, but I can't touch the thin vertical "center to gps"
bar because it is a thin vertical area in the extreme right side of the
screen; thus gpsmid always takes the touch as a zoom in or out."