Use mainStreetNet in route calculations to make large routes possible to be calculated on the mobile:
Actually I've just calculated a route with 155km from somewhere North of Munich to somewhere South of Munich. GpsMid mainly used the motorways and primary roads to surrund Munich and the route calculation increased the heap space of my mobile to only 5.5 MB even though deep route calculation was used and turn restrictions were activated.

MainStreetNet calculation is used only for routeModes that have the mainStreetNet flag set in the style-file. Also there's a new Routing Option that determines how far away from routeStart and RouteTarget GpsMid will only load and examine connections in the mainStreetNet. (If the setting would lead to "No solution" GpsMid will retry the calculation a few times with starting to try the mainstreetNet on a larger distance)

Also GpsMid's route calculation will only start to use the mainStreetNet if at least 20 connections were already examined by the routeCalculation in non-MainstreetNode.

Hope this does not break anything!!! But if it does in some special cases set mainStreetNet to false in the style-file or set in the Routing Options the mainStreetNet distance to 99999 for disabling this functionality.