Testing out what it would take to support I18N / L10N
This commit is a test to see what it would take to translate GpsMid into different
languages and if it is feasible.

This commit currently only enables translation of two strings in the splash-screen

Please let me know what you think of the idea of trying to I18N GpsMid.

You can find some documentation about the I18N framework at
http://www.j2mepolish.org/cms/leftsection/documentation/localization.html