[patch #2872401] a patch to use http://www.rss2html.com/ for opening geonames
Patch thanks to jkpl

Online Info" offers a Wikipedia service, which is a RSS feed. Nokia
phones (well, at least the S60 E71, but I think also others) don't appear
to be very smart about handling RSS feeds when opened in the browser
in the way by gpsmid. The phone adds the RSS feed to a RSS feed selection
the first time, and
after that it just complains the feed is already there, and doesn't open
the RSS
feed, instead just shows an empty screen.