Take traffic signals into account for ETA calculation:
RouteNodes that are at maximum 24m away from traffic signals are marked by Osm2GpsMid (and are shown in GpsMid if show route connections are activated).
For ETA calculation a penalty of 20 secs is added to the route duration, if the last traffic lights routeNode was at least 200 meters before the current one.

Traffic Lights are not taken into account for the route calculation itself.