The fire button in IconMenus didn't work on Nokia phones
On Nokia phones (S60 S40), the fire button is used to open up the
list of commands in fullscreen mode as soon as there is a single
Command set. They also don't have a back button for which the
back command is important. Thus, don't add the back command in fullscreen menu icons
on these phones.

It uses a CFGBIT to decide to include the backcommand, so it could
be made configurable or extended to other devices.