[ patch 2761048 ] j2me zipfile reading patch (no permission nags anymore)
This patch was provided by cmuelle8 / trendypack. Many thanks for the very helpful patch for this much needed
functionality

With this patch it is now possible to read the map data from a zip file and thus separate out the application jar and
the mapdata.

Unfortunately the patch still has a bunch of problems. It still does not start at all on my Nokia 6220c (S60) which
seems to be an issue with the underlying ZIP implementation taken from j2mepolish. It quickly runs out of memory on my
Nokia 6301 (S40) using a 6Mb mapfile. It does however seem to work on my SE K550 without problems. I have also had this
patch in my local tree for several months, so I am confident it doesn't cause any harm when the zip reading is not
activated. On the otherhand it would be good to get more exposure to see under what conditions and on what mobiles it
might already work. It is also hopefully easier to improve on this patch once it is in the tree.