Produce the route line in a separate thread:
This allows especially after calculating long routes to give the first instructions almost immediately while the rest of the route line and instructions is determined in the background. During route line production GpsMid counts up the distance from the current position to the end of the already determined route line part in the route distance field with the background color. ETA and duration are not shown during route line production.

It is now also possible to abort the route line production by stopping the route calculation.