Add a waypoint on OSM POI uploads
As we don't update our data when editing OSM, there is little feedback to the user that editing was successfull.
In particular for adding POIs the user can't check what and where they uploaded the POI.

As an interims solution until we find a solution to update the GpsMid with the uploaded data, add a waypoint
for every added OSM POI to give feedback to the user. Not ideal, but better than nothing.