[build fix] Generic/minimal was broken, as Trace was calling the Noisemaker.stopPlayer
This is a quick fix and probably not the right one. Also NoiseMaker seems like a big
mess with respect to the //#if polish.api.mmapi   Only parts of the code are hidden behind the pre processor
directive. I presume though that without mmapi none of the sound works, so we might as well
not include any code in this case. Can someone check this?