Make nearest POI search POI type selection menu "searchable"
So far the the POI type selection menu was a simple list and one could only linearly
move through it to select the POI type for which to search. With more and more POI types
included in the default style, this has become a bit of a pain. So instead use the same
interface as used in the POI type selection to add a new POI to OSM in the GpsMid editing
support and similar to the main search.

One problem with it though is that I had to drop the maximum distance selection edit field,
as the new interface doesn't support any additional UI elements. I thought the gain in
usability to select the POI type was worth the loss of max radius choice, but if others
disagree, please let me know.