[BugFix #2884988] Osm2GpsMid GUI can't just disappear
When Osm2GpsMid is started from the gui, the text output to the console is redirected
to a new "Console output" TextArea at the bottom of the GUI.

This should fix the issue that if you start Osm2GpsMid by double clicking on the jar,
it just dissapears while converting. Not the most elegant solution but should be just about
doable and can be improved upon later.