[ OSM-editing ] Add an address-node collector mode
I often find adding addresses to OSM rather tedious, yet they are very important.
Try and make adding address nodes easier by adding a new Add-address-node mode to GpsMid.
It is a simple form that allows you to enter very easily house number, house name, postcode and street name
and directly upload it to OSM.

As with the other editing features, this currently needs internet connectivity. It is also necessary
to be careful to use uptodate maps which include displaying of addr:housenumber tags, to ensure
risk of adding duplicate data is low.