In Wizard mode, also output info and errors to console
Recently, the output and error stream were redirected to the GUI,
as some people might start the GUI via double click and can't see
the console. However, sometimes the GUI might close too early and
it can be hard to see all the errors. So now the info and error
messages get printed both on the console and in the GUI