Add another style-file with a reduced set of polygone definitions
With more and more imports of buildings and landuse areas happening around the world, the data quantity of these
features can easily swamp out that of other features significantly increasing the amount of resources needed to cope
with these larger midlets. As these features are imho of somewhat limited use in navigation, this style is basically the
same as the normal one, but without landuse areas and building polygons and will be used in the largest of the "ready-to-go" maps offered.