bugfix: AreaPOIs were getting assigned to multiple nodes of the way.
When the type of a way gets calculated and it is determined that it is an area POI, the type gets coppied to the first node
without a type. However, if this function is called again, it will now copy it to the next node, as the first node now has a type.
Thus this function is only allowed to be called once. There was already a check to prevent this, but it was triggered on the way having
a type assigned. However, as it is an areaPOI, there is no way type, only a node type, so the check failed.

This caused all of the nodes of the area way to get assigned POIs rather than just one and caused the "Couldn't not
assign area POI, as no node without type" warning.