Optimisation: Only store tags that are needed according to the style-file
This builds up a list of relevant keys in LegendParser and then only includes key-value
pairs during parsing if the key is in the relevant key list. This should hopefully save some
memory for all the keys that are in the data, but never used by Osm2GpsMid.

WARNING: Please double check that all of the relevant keys correctly added and I didn't miss any, as
otherwise things will break.