Optimisation: At the end of parsing, replace the HashMap of Nodes with a simple Vector of Nodes to save some memory
The HashMap implementation has more overhead per element than a simple Vector in order to allow for fast and efficient
lookup by key.  This is necessary during initial parsing, where for each Way node references, we have to look up by key
if the node exists. After parsing, we only appear to access it sequentially, so replace the HashMap with a Vector to save
some memory during later stages of processing.