(sortof) memory optimisation: Free some memory after it is no longer used by setting variables to null
Also remove a static declaration on ways, as otherwise that memory was never released, as there was still
the static reference even after the Object was freed.

This freeing happens right at the end, so it is only of limited use. There is still the jar creation afterwards
though. However, given that the JavaVM never releases memory back to the OS, it just ends up increasing the
free memory available to Java which at this stage is kind of useless. :-S