memory optimization: Free all unused ways (and their nodes) at the end of parsing.
So far, all ways that were in the bounding box were retain, irrespective of if they were relevant in the style-file.
Now, these ways get deleted again at the end of parsing and after triangulation, where empty ways might be used to
define the polygon. Depending on style-file, this can save a considerable amount of memory, especially, as this also
frees quite a few then unused nodes.