[bug fix] too many ways error message does not show up
The index to count the number of ways and pos was itself only a byte. So by definition it can't be bigger than
what fits in a byte which is what this error message is supposed to detect. So change the internal counter to int
to be able to detect overflow.

(Compile tested, but not actually verified that it works)