Initial implementation of the new OSM Google Protobuf binary format.
Osm2GpsMid should now support the binary .osm.pbf format as well as the
standard xml based .osm format.

The .osm.pbf format is considerably faster in parsing, especially faster than the
decompression of the .bz2 files.

It seems to work in that the filesizes are of the same size, but it isn't fully tested yet, so use with care!