Change by karel_vasicek: Regarging "pausing GpsMid" - I changed line 762 in Trace.java to: if (locationProducer != null && !gpx.isRecordingTrk()) {
This hopefully solves [ gpsmid-Bugs-3483706 ] blackberry disconnects from gps when run in background and billi25's issue: GpsMid
for Android does not record the track (a) after the screen is off,
(b) if GpsMid is put to background.

This change does not seem to make any difference on S60V5 (Nokia 5800), it seems GpsMid records with and without the change track points in the background.